owd: an output watchdog
owd is an "output watchdog" which invokes a subprocess and kills it
if it fails to produce output (via stdout or stderr) for more than
a specified amount of time.  owd properly handles Unix process
semantics including signals and process groups.  By default it looks
for any output, but it can optionally insist on matching a specified
regular expression.